+++ engine/Runtime/RenderPipeline/Handler/CreateVAOHandler.h
@@ -21,7 +21,7 @@ namespace DivineBrush {
         unsigned int vboHandle = 0;
         unsigned int vertexDataSize = 0;
         void *vertexData = nullptr;
-        unsigned int vertexIndexDataSize = 0;
+        unsigned short vertexIndexDataSize = 0;
         void *vertexIndexData = nullptr;
 
         void Run() override;
